pyproject.toml
[project]
name = "multi-agent-system"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "langgraph (>=1.0.4,<2.0.0)",
    "langchain (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "numpy (>=2.3.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "pymupdf (>=1.26.6,<2.0.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
